refactor: BlackJackManager 패키지 domain 밑으로 이동
