+++ src/main/java/blackjack/domain/BlackJackManager.java
@@ -1,4 +1,4 @@
-package blackjack.domain.gamer;
+package blackjack.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -8,6 +8,8 @@ import java.util.stream.Collectors;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 import blackjack.domain.card.Deck;
+import blackjack.domain.gamer.Dealer;
+import blackjack.domain.gamer.Player;
 
 public class BlackJackManager {
